int8_t

+++ components/tc_bus/tc_bus.cpp
@@ -152,10 +152,10 @@ namespace esphome
 
             #ifdef USE_API
 
-            auto api_userservicetrigger = new api::UserServiceTrigger<uint32_t>("on_acknowledge", {"type"});
+            auto api_userservicetrigger = new api::UserServiceTrigger<int8_t>("on_acknowledge", {"type"});
             api::global_api_server->register_user_service(api_userservicetrigger);
-            auto automation = new Automation<uint32_t>(api_userservicetrigger);
-            auto lambdaaction = new LambdaAction<uint32_t>([this](uint32_t type) -> void {
+            auto automation = new Automation<int8_t>(api_userservicetrigger);
+            auto lambdaaction = new LambdaAction<int8_t>([this](int8_t type) -> void {
                 // Something
             });
             automation->add_actions({lambdaaction});
